+++ packages/rax-icon/src/miniapp/index.ts
@@ -1,7 +1,7 @@
 import fmtEvent from './fmtEvent';
 
 const noop = () => {};
-
+console.warn('组件所依赖的 rax-icon 版本较旧，请尽快重新构建发布该组件');
 Component({
   data: {
     styleSheet: '',