+++ packages/rax-scrollview/src/weex/index.tsx
@@ -12,6 +12,7 @@ import RefreshControl from 'rax-refreshcontrol';
 import getElementById from 'rax-get-element-by-id';
 import cx from 'classnames';
 import { ScrollViewProps } from '../types';
+import wrapDefaultProperties from '../utils/wrapDefaultProperties';
 import '../index.css';
 
 const baseCls = 'rax-scrollview';
@@ -179,4 +180,4 @@ const ScrollView: ForwardRefExoticComponent<ScrollViewProps> = forwardRef(
   }
 );
 
-export default ScrollView;
+export default wrapDefaultProperties(ScrollView);