+++ packages/rax-slider/src/miniapp/index.ts
@@ -1,6 +1,7 @@
 import fmtEvent from './fmtEvent';
 
 const noop = () => {};
+console.warn('组件所依赖的 rax-slider 版本较旧，请尽快重新构建发布该组件');
 Component({
   data: {
     current: 0,