+++ packages/rax-textinput/src/index.tsx
@@ -8,7 +8,7 @@ import {
   useState,
   useMemo
 } from 'rax';
-import { isWeex, isWeb, isWeChatMiniProgram, isNode, isMiniApp } from 'universal-env';
+import { isWeex, isWeb, isNode, isMiniApp } from 'universal-env';
 import setNativeProps from 'rax-set-native-props';
 import keyboardTypeMap from './keyboardTypeMap';
 import {
@@ -131,15 +131,8 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
       autoComplete: autoComplete && 'on',
       maxlength: maxlength || maxLength,
       onChange: (onChange || onChangeText) && handleChange,
-      onInput: (e: InputEvent) => {
-        onInput && handleInput(e);
-        if (!isWeChatMiniProgram) {
-          forceUpdate(tick => tick + 1);
-        }
-      },
       onBlur: onBlur && handleBlur,
-      onFocus: onFocus && handleFocus,
-      ref: refEl
+      onFocus: onFocus && handleFocus
     };
 
     // Diff with web readonly attr, `disabled` must be boolean value
@@ -178,6 +171,7 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
           }` }} />
           <textarea
             {...propsCommon}
+            ref={refEl as Rax.MutableRefObject<HTMLTextAreaElement>}
             className={['rax-textinput', styleClassName, className || ''].join(' ')}
             style={{
               ...style,
@@ -190,6 +184,12 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
             value={value}
             confirm-type={confirmType}
             show-count={showCount}
+            onInput={(e: Rax.ChangeEvent<HTMLTextAreaElement>) => {
+              onInput && handleInput(e);
+              if (isMiniApp) {
+                forceUpdate(tick => tick + 1);
+              }
+            }}
           >
             {/* undefined will be rendered to comment node in ssr */}
             {!isWeex && (propsCommon.value || defaultValue || '')}
@@ -204,6 +204,7 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
           }` }} />
           <input
             {...propsCommon}
+            ref={refEl as Rax.MutableRefObject<HTMLInputElement>}
             className={['rax-textinput', styleClassName, className || ''].join(' ')}
             style={{
               ...style,
@@ -216,6 +217,12 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
             random-Number={randomNumber}
             selection-start={selectionStart}
             selection-end={selectionEnd}
+            onInput={(e: Rax.ChangeEvent<HTMLInputElement>) => {
+              onInput && handleInput(e);
+              if (isMiniApp) {
+                forceUpdate(tick => tick + 1);
+              }
+            }}
           />
         </Fragment>
       );