+++ packages/rax-video/src/miniapp/index.ts
@@ -34,6 +34,7 @@ const componentMethods = supportMethods.reduce((prev, current) => {
   return prev;
 }, {});
 
+console.warn('组件所依赖的 rax-video 版本较旧，请尽快重新构建发布该组件');
 Component({
   data: {},
   props: {