+++ packages/rax-icon/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 1.4.3
+
+- Fix types export bug
+
+## 1.4.2
+
+- Keep wechat native code in older dir
+- Remove `sideEffects: false` in package.json
+
+## 1.4.1
+
+- Keep miniapp native code in older dir
+- Add `sideEffects: false` in package.json
+- Remove `quickappConfig` in package.json
+
+## 1.4.0
+
+- Support exports field
+- Support more miniapp platforms
+- Reduce npm files
+- Adjust miniapp native code
+
 ## v1.3.8
 
 - can't load fontface after first loading