+++ packages/rax-modal/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 1.6.5
+
+- Keep wechat native code in older dir
+
+## 1.6.4
+
+- Remove `quickappConfig` in package.json
+
+## 1.6.3
+
+- Keep miniapp native code in older dir
+
+## 1.6.2
+
+- Reduce npm files
+- Adjust miniapp native code
+
 ## v1.6.1
 
 - fix overflow in android web
@@ -7,8 +24,8 @@
 ## v1.6.0
 
 - update dependencies:
-  - rax-text: v1.x to v2.x
-  - rax-view: v1.x to v2.x
+- rax-text: v1.x to v2.x
+- rax-view: v1.x to v2.x
   
 ## v1.5.6