+++ packages/rax-scrollview/CHANGELOG.md
@@ -1,15 +1,38 @@
 # Changelog
 
-## [3.4.2] - 2021-03-29
+## 3.6.1
+- Fix `pixelRatio` is not initialized and make onEndReached can not triggered
+
+## 3.6.0
+
+- Support `string` type in runtime miniapp and web, of onEndReachedThreshold, scrollTo
+
+## 3.5.2
+
+- Remove `sideEffects: false` in package.json
+
+## 3.5.1
+
+- Keep miniapp native code in older dir
+- Add `sideEffects: false` in package.json
+- Remove `quickappConfig` in package.json
+
+## 3.5.0
+
+- Support exports field
+- Support more miniapp platforms
+- Adjust miniapp native code
+
+## 3.4.2
 
 - Format npm output files
 
-## [3.4.1] - 2021-03-23
+## 3.4.1
 
 ### Added
 
 - Set `enable-flex` in miniapp to true to support horizontal scroll-view in wechat miniprogram
-## [3.4.0] - 2021-03-16
+## 3.4.0
 
 ### Added