+++ packages/rax-waterfall/CHANGELOG.md
@@ -1,1 +1,14 @@
 # Changelog
+
+## 1.2.2
+
+- Remove `sideEffects: false` in package.json
+
+## 1.2.1
+
+- Add `sideEffects: false` in package.json
+
+## 1.2.0
+
+- Support more miniapp platforms
+- Reduce npm files